requirements.txt
anthropic>=0.18.0
rich>=13.0.0
typer>=0.9.0
pyyaml>=6.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
